refactor: The update section of packages now have more generic fields
Instead of `github` and `index`. The fields are now `version` and
`download`. The `version` is a url where the version should be obtained.
`download` is an optional url for where the download url can be
obtained. If no `download` is specified, it will be inferred from the
version url if possible.